Move loop op lowering from NN/Conv.cpp to ControlFlow/Loop.cpp.
