+++ src/Conversion/ONNXToKrnl/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(OMONNXToKrnl
         ONNXToKrnlCommon.cpp
         ONNXToKrnlCommon.hpp
+        ControlFlow/Loop.cpp
         Math/Elementwise.cpp
         Math/Gemm.cpp
         Math/MatMul.cpp
